wait Long/Integer implemented

+++ eReg/eRegulation/Thread_Thermometers.java
@@ -36,12 +36,12 @@ public class Thread_Thermometers implements Runnable
 			LogIt.tempData();
 			
 			Long timeElapsed					= Global.now() - timeStart;
-			Integer timeIncrement 				= 10000 - timeElapsed.intValue();
+			Long timeIncrement 					= 10000L - timeElapsed;
 			
 			if (timeIncrement <= 0)
 			{
 				LogIt.error("Thread_Thermometers", "run", "Temp reads over 10 seconds");
-				timeIncrement					= 1000;
+				timeIncrement					= 1000L;
 			}
 
 			Global.waitMilliSeconds(timeIncrement);
